backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
slowapi==0.1.9
email-validator==2.2.0
ruff==0.8.4
black==24.10.0
piper-tts==1.2.0
onnxruntime==1.20.1
kokoro==0.3.4
misaki[en]==0.6.7
soundfile==0.12.1
numpy==1.26.4
